--- CPP_Files/Pickup.h
@@ -1,36 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "GameFramework/Actor.h"
-#include "Pickup.generated.h"
-
-UCLASS()
-class WORKSHEET_C_API APickup : public AActor
-{
-	GENERATED_BODY()
-	
-public:	
-	// Sets default values for this actor's properties
-	APickup();
-
-	// Called when the game starts or when spawned
-	virtual void BeginPlay() override;
-	
-	// Called every frame
-	virtual void Tick( float DeltaSeconds ) override;
-
-	UPROPERTY(EditAnywhere)
-	USceneComponent* PickupRoot;
-
-	// Static mesh for the pickups
-	UPROPERTY(EditAnywhere)
-	UStaticMeshComponent* PickupMesh;
-
-	UPROPERTY(EditAnywhere)
-	UShapeComponent* PickupBox;
-
-	UFUNCTION()
-	void OnPlayerEnterPickupBox(UPrimitiveComponent* OverlappedComp, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult&SweepResult);
-
-};
