+++ worksheet_C/CPP_Files/Pickup.cpp
@@ -1,0 +1,44 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "Worksheet_C.h"
+#include "Pickup.h"
+
+
+// Sets default values
+APickup::APickup()
+{
+ 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
+	PrimaryActorTick.bCanEverTick = true;
+
+	PickupRoot = CreateDefaultSubobject<USceneComponent>(TEXT("PickupRoot"));
+	RootComponent = PickupRoot;
+
+	PickupMesh = CreateDefaultSubobject<UStaticMeshComponent>(TEXT("PickupMesh"));
+	PickupMesh->AttachToComponent(PickupRoot, FAttachmentTransformRules::SnapToTargetNotIncludingScale);
+
+	PickupBox = CreateDefaultSubobject<UBoxComponent>(TEXT("PickupBox"));
+	PickupBox->SetWorldScale3D(FVector(1.0f, 1.0f, 1.0f));
+	PickupBox->bGenerateOverlapEvents = true;
+	PickupBox->OnComponentBeginOverlap.AddDynamic(this, &APickup::OnPlayerEnterPickupBox);
+	PickupBox->AttachToComponent(PickupRoot, FAttachmentTransformRules::SnapToTargetNotIncludingScale);
+}
+
+// Called when the game starts or when spawned
+void APickup::BeginPlay()
+{
+	Super::BeginPlay();
+	
+}
+
+// Called every frame
+void APickup::Tick( float DeltaTime )
+{
+	Super::Tick( DeltaTime );
+
+}
+
+void APickup::OnPlayerEnterPickupBox(UPrimitiveComponent * OverlappedComp, AActor * OtherActor, UPrimitiveComponent * OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult & SweepResult)
+{
+	Destroy();
+}
+
