--- worksheet_E/README.md
@@ -1,1 +1,0 @@
-# Worksheet E
